refactor: move Minimap code to Minimap
