+++ src/main/java/org/runejs/client/RSString.java
@@ -8,7 +8,6 @@ import java.awt.*;
 import java.nio.charset.StandardCharsets;
 
 public class RSString {
-    public static ProducingGraphicsBuffer mapbackProducingGraphicsBuffer;
     public static CacheArchive aCacheArchive_1705;
     public static int[] keyCodes = new int[128];
     public static int tooltipDelay = 50;
